+++ ext/stub/45__closure.zep.c
@@ -12,6 +12,7 @@
 #include <Zend/zend_interfaces.h>
 
 #include "kernel/main.h"
+#include "kernel/operators.h"
 #include "kernel/object.h"
 #include "kernel/memory.h"
 
@@ -20,23 +21,25 @@ ZEPHIR_INIT_CLASS(stub_45__closure)
 {
 	ZEPHIR_REGISTER_CLASS(stub, 45__closure, stub, 45__closure, stub_45__closure_method_entry, ZEND_ACC_FINAL_CLASS);
 
-	zend_declare_property_null(stub_45__closure_ce, SL("x"), ZEND_ACC_PUBLIC);
+	zend_declare_property_null(stub_45__closure_ce, SL("counter"), ZEND_ACC_PUBLIC);
 	return SUCCESS;
 }
 
 PHP_METHOD(stub_45__closure, __invoke)
 {
+	zend_long _0;
 	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
-	zval x;
+	zval counter;
 	zval *this_ptr = getThis();
 
-	ZVAL_UNDEF(&x);
+	ZVAL_UNDEF(&counter);
 	ZEPHIR_METHOD_GLOBALS_PTR = pecalloc(1, sizeof(zephir_method_globals), 0);
 	zephir_memory_grow_stack(ZEPHIR_METHOD_GLOBALS_PTR, __func__);
-	zephir_memory_observe(&x);
-	zephir_read_property(&x, this_ptr, SL("x"), PH_NOISY_CC);
+	zephir_read_property(&counter, this_ptr, SL("counter"), PH_NOISY_CC | PH_READONLY);
 
-	RETVAL_ZVAL(&x, 1, 0);
+	_0 = (zephir_get_numberval(Z_REFVAL_P(&counter)) + 1);
+	ZVAL_LONG(Z_REFVAL_P(&counter), _0);
+	RETVAL_ZVAL(Z_REFVAL_P(&counter), 1, 0);
 	RETURN_MM();
 }