+++ ext/stub/50__closure.zep.h
@@ -1,0 +1,14 @@
+
+extern zend_class_entry *stub_50__closure_ce;
+
+ZEPHIR_INIT_CLASS(stub_50__closure);
+
+PHP_METHOD(stub_50__closure, __invoke);
+
+ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_50__closure___invoke, 0, 0, 0)
+ZEND_END_ARG_INFO()
+
+ZEPHIR_INIT_FUNCS(stub_50__closure_method_entry) {
+PHP_ME(stub_50__closure, __invoke, arginfo_stub_50__closure___invoke, ZEND_ACC_PUBLIC|ZEND_ACC_FINAL)
+	PHP_FE_END
+};